parser: stop parsing the indented variable blocks if no more indentation found
This change requires to `peek` the next token as the `parse_indented_variable_block()` function does not consume the next token if it is exiting, allowing the subsequent calls to the lexer start from that token that triggered the exit.

The logos::Lexer does not implement peeking and so a wrapper was introduced and a translation between TokenKind and Token because the logos Token also doesn't implement cloning and so it's not possible to store its value.