refactor: Allow for dead-code in graph.rs
There are a few functions that aren't used in the main application just
yet, while they have some coverage in the unit testing. Going forward,
I would like to avoid this by doing testing the other way around, but
for now, I would like this code to be green in all ways.