feat(lexer): Deal with assingments
Having a separate '=' token is not doing me any favours because there is
no way to differentiate between Identifier tokens and whatever the
assignment value would be.

Since I know that the assignment values are _always_ going to start with
an equal sign, this is an easy way to know that the rest of the string
will be a value that I don't have to do anything here.